requirements.txt
beautifulsoup4==4.11.1
Brotli==1.0.9
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
numpy==1.23.1
opencv-python==4.6.0.66
pandas==1.4.3
Pillow==9.2.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
requests-tor==1.3
six==1.16.0
soupsieve==2.3.2.post1
stem==1.8.0
urllib3==1.26.11
